backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Async database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0

# HTTP client (async — never use requests)
httpx>=0.27.0

# AWS S3 (async — never use boto3 directly)
aioboto3>=13.0.0

# LLM providers
anthropic>=0.49.0
openai>=1.30.0

# ElevenLabs (stretch goal)
elevenlabs>=1.4.0

# WebSockets
websockets>=12.0

# Skribby listener needs websockets (above)
